internal/driver: Move VolumeLifecycle to its own file
The type was defined in controllerserver.go, but was not actually used
there. It's used in nodeserver.go and luks.go, but it's barely used.
Nothing actually sets it, as far as I can tell. Maybe it should be
removed...
